+++ crates/block/hadris-fat/src/async_send.rs
@@ -6,8 +6,11 @@ macro_rules! io_transform {
 use hadris_storage::async_send as storage;
 
 macro_rules! impl_fat_driver {
-    (impl[D: BlockDevice, T: NodeTable] $($rest:tt)*) => {
-        hadris_fs::impl_fs_driver!(async_send, impl[D: BlockDevice, T: NodeTable<With<Node>: Send>] $($rest)*);
+    (impl[D: BlockDevice, T: NodeTable, C: Clock, P: CodePage] $($rest:tt)*) => {
+        hadris_fs::impl_fs_driver!(
+            async_send,
+            impl[D: BlockDevice, T: NodeTable<With<Node>: Send>, C: Clock + Send, P: CodePage + Send] $($rest)*
+        );
     };
 }